Role Based Authentication And Security Also Protect APIS Using getServerSession

+++ app/api/product/[slug]/route.ts
@@ -6,6 +6,8 @@ import { NextRequest, NextResponse as res } from "next/server"
 import ServerCatchError from "@/lib/server-catch-error"
 import ProductModel from "@/models/product.model"
 import SlugInterface from "@/interface/slug.interface"
+import { getServerSession } from "next-auth"
+import { authOptions } from "../../auth/[...nextauth]/route"
 
 export const GET = async (req: NextRequest, context: SlugInterface) => {
     try {
@@ -24,6 +26,13 @@ export const GET = async (req: NextRequest, context: SlugInterface) => {
 
 export const PUT = async (req: NextRequest, context: SlugInterface) => {
     try {
+        const session = await getServerSession(authOptions)
+        if(!session)
+            return res.json({message: "Unauthorized"}, {status: 401})
+
+        if(session.user.role !== "admin")
+            return res.json({message: "Unauthorized"}, {status: 401})
+
         const {slug: id} = context.params
         const body = await req.json()
         const product = await ProductModel.findByIdAndUpdate(id, body, {new: true})
@@ -40,6 +49,13 @@ export const PUT = async (req: NextRequest, context: SlugInterface) => {
 
 export const DELETE = async (req: NextRequest, context: SlugInterface) => {
     try {
+        const session = await getServerSession(authOptions)
+        if(!session)
+            return res.json({message: "Unauthorized"}, {status: 401})
+
+        if(session.user.role !== "admin")
+            return res.json({message: "Unauthorized"}, {status: 401})
+        
         const {slug: id} = await context.params
         const product = await ProductModel.findByIdAndDelete(id)
